Refactor - Moved all queries into a single file
