+++ src/components/views/home/Home.js
@@ -5,49 +5,10 @@ import PortfolioWidget from "./../../widgets/portfolioWidget/PortfolioWidget";
 import EducationWidget from "./../../widgets/educationWidget/EducationWidget";
 import ResumeWidget from "./../../widgets/resumeWidget/ResumeWidget";
 import { useContentful } from "./../../../Hooks/useContentful";
-import { gql } from "@apollo/client";
-
-const GQL_QUERY = gql`
-	query GetHomeInfo($language: String!) {
-		aboutMeCollection(locale: $language) {
-			items {
-				title
-				description
-			}
-		}
-		skillsCollection {
-			items {
-				sys {
-					id
-				}
-				title
-				skillList
-			}
-		}
-		portfolioCollection(locale: $language, limit: 2) {
-			items {
-				sys {
-					id
-				}
-				title
-				tumbnail {
-					url
-				}
-			}
-		}
-		educationCollection(locale: $language, limit: 3) {
-			items {
-				subtitle
-				image {
-					url
-				}
-			}
-		}
-	}
-`;
+import { HOME_QUERY } from "./../../../graphql/queries";
 
 function Home() {
-	const [content, ,] = useContentful(GQL_QUERY);
+	const [content, ,] = useContentful(HOME_QUERY);
 
 	return (
 		<>
