+++ src/index.js
@@ -5,13 +5,7 @@ import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import { ApolloProvider } from "@apollo/client";
 import WithTheme from "./WithTheme";
-import { CLIENT } from "./Hooks/apollo";
-
-// const CLIENT = new ApolloClient({
-// 	uri:
-// 		"https://graphql.contentful.com/content/v1/spaces/ui8qz5ptyq23?access_token =VRsqmQ37waPrLOGCNmmwigcY2mekXcIA9Q-08kSRDkE",
-// 	cache: new InMemoryCache(),
-// });
+import { CLIENT } from "./graphql/apollo";
 
 const AppWithTheme = WithTheme(App);
